rename app_state_socialnav.py to app_state_socialnav_study.py
